+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # MMM-Fuel Changelog
 
+## [2.4.0]
+
+### Added
+
+* [Provider: `gasbuddy`](https://github.com/fewieden/MMM-Fuel/pull/97)
+* [Config option: `fade`](https://github.com/fewieden/MMM-Fuel/issues/7)
+* [Config option: `showAddressCity`](https://github.com/fewieden/MMM-Fuel/pull/64)
+
+### Changed
+
+* [Migrated provider `nsw` from API v1 to v2 in order to support new region Tasmania (Australia)](https://github.com/fewieden/MMM-Fuel/issues/68#issuecomment-916505199)
+* Dependency update
+
 ## [2.3.1]
 
 ### Fixed